requirements.txt
# Dependencias principales
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
sahi>=0.11.0

# Procesamiento de imágenes
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Visualización y análisis
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilidades
tqdm>=4.65.0
PyYAML>=6.0
pandas>=2.0.0
scipy>=1.10.0
numpy>=1.24.0

# Herramientas opcionales (descomentar si necesitas)
# labelImg  # Para anotación manual
# roboflow  # Para gestión de datasets
# wandb     # Para tracking de experimentos
# tensorboard  # Para visualización de métricas
